requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.21.0
pillow>=9.0.0
tensorflow>=2.13.0,<2.18.0
keras>=2.13.0
protobuf<4.21.0